+++ CHANGELOG.md
@@ -56,11 +56,10 @@
 
 ## 1.2.0
 
-- Database backend converted to use SQLAlchemy
+- Database backend converted to use SQLAlchemy (2.0)
 - Complete test suite created
 - Banking interface updated to use subtransactions
 - Added the ability to transfer statements when a new credit card is added to an account
 - Fixed error on handling subtransactions in credit forms
 - Fixed error updating transaction display from widget bar
-- Adding minimal set of tests