+++ CHANGELOG.md
@@ -64,10 +64,11 @@
 - Fixed error updating transaction display from widget bar
 
 
-### 1.2.1 (in-progress)
+### 1.2.1
 
 - Improve development functionality; better/safer cleaning with a separate database and preloaded data
 - Add custom form fields for consistent app-wide form implementations
+- Overhaul documentation formatting with _isort_ and _Black_
 - Fixed bug where multiple subtransactions failed to update properly
 - Fixed error where validation was not raised for blank inputs to a `CustomChoiceSelectField`
 - Fixed errors in field validation for nested subforms with unset values